Chore: Move fund service code to specified dir
