+++ src/test/java/com/hana/fund/SelectTests.java
@@ -1,7 +1,7 @@
 package com.hana.fund;
 
 import com.hana.app.data.entity.fund.Fund;
-import com.hana.app.service.FundService;
+import com.hana.app.service.fund.FundService;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
@@ -21,6 +21,7 @@ public class SelectTests {
     void contextLoads() {
         List<Fund> x = fundService.getAll();
         for (Fund fund : x) {
+            log.info(fund.getId().toString());
             log.info(fund.getName());
             log.info(fund.getRiskType().toString());
         }
